Split interpreter.rs into eval/ submodule
Move statement execution and expression evaluation out of the 410-line
interpreter.rs into eval/{statements,expressions}.rs as free functions
that take &mut Interpreter. interpreter.rs now holds only the public
API, the struct definition, and the still-monolithic module loader
(which moves in the next step).

Struct fields are pub(crate) so the eval submodule can mutate them
directly — simpler than adding narrow accessors for a hobby project.
