+++ crates/amarok_interpreter/src/eval/mod.rs
@@ -1,0 +1,9 @@
+//! Free-function execution logic that operates on `&mut Interpreter`.
+//!
+//! Split out from `interpreter.rs` so that file can stay focused on the
+//! public API. The submodules here hold the recursive AST walk:
+//! - `statements` — statement-level execution and control flow
+//! - `expressions` — expression evaluation and function dispatch
+
+pub(crate) mod expressions;
+pub(crate) mod statements;
